// models/SqlDB.ts
export type DBColumnType =
  | 'bigint'
  | 'binary'
  | 'bit'
  | 'char'
  | 'date'
  | 'datetime'
  | 'datetime2'
  | 'datetimeoffset'
  | 'decimal'
  | 'FILESTREAM'
  | 'float'
  | 'image'
  | 'int'
  | 'money'
  | 'nchar'
  | 'ntext'
  | 'numeric'
  | 'nvarchar'
  | 'real'
  | 'rowversion'
  | 'smalldatetime'
  | 'smallint'
  | 'smallmoney'
  | 'sql_variant'
  | 'text'
  | 'time'
  | 'timestamp'
  | 'tinyint'
  | 'uniqueidentifier'
  | 'varbinary'
  | 'varchar'
  | 'xml';